LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-iframe.html

<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<p>This test loads a secure iframe that loads an insecure iframe.  We should
get a mixed content callback becase the secure inner frame should block mixed content.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-iframe.html";
></iframe>
</body>
</html>
