LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-iframe-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-iframe.html requested insecure content from http://127.0.0.1:8080/security/mixedContent/resources/boring.html. This content was blocked and must be served over HTTPS.

PASS window.data is 'blocked'
This test loads a secure iframe that loads an insecure iframe. We should get a mixed content callback becase the secure inner frame should block mixed content.
